Add feature/rule ValidateOcdidLowerCase
Added a feature that validates if the ocdids in the XML feed are in all lowercase. Throws out a warning message otherwise.
Added a sample file ocdid_lowercase_sample.xml to test out the feature - ValidateOcdidLowerCase. It should have 2 warning messages for the inappropriate ocdids (intentionally added) in the feed.