Update the validator to work with the new EU OCD ID format and deprecate check_github (-g) flag.
All EU OCD IDs are no longer contained within one file but have been disbanded to the individual country files. Therefore all EU country files need to be cached when validating an EU feed. Corresponding GitHub commit and fingerprint checks have been removed for a simpler caching experience. Individual cached files older than an hour will be redownloaded via the raw GitHub URL (not an API call).

Due to the removal of the Github API calls, the check_github flag (passed in as -g) is being deprecated. The flag itself is not being removed within these changes but is no longer considered as part of the OCD ID logic for refreshing the local cache. Full removal of the flag will be completed in a future release.

PiperOrigin-RevId: 724336529